requirements.txt
django == 5.1.5
python-dotenv == 1.0.1